requirements.txt
aiohttp
aiosqlite
discord.py==2.6.3
python-dotenv
mcstatus
aiomcrcon
